requirements.txt
uuid
loguru
python-socketio
aioreloader
aioharmony
tellcore-py